Add a state holder to ProductFormScreen.
Add a state full composable to ProductFormScreen.
