+++ app/src/main/java/com/alura/aluvery/ui/activities/ProductFormActivity.kt
@@ -3,28 +3,10 @@ package com.alura.aluvery.ui.activities
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.*
-import androidx.compose.runtime.*
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardCapitalization
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import coil.compose.AsyncImage
-import com.alura.aluvery.R
+import androidx.compose.material.Surface
 import com.alura.aluvery.dao.ProductDao
-import com.alura.aluvery.model.Product
+import com.alura.aluvery.ui.screens.ProductFormScreen
 import com.alura.aluvery.ui.theme.AluveryTheme
-import java.math.BigDecimal
 
 class ProductFormActivity : ComponentActivity() {
     private val dao = ProductDao()
@@ -33,167 +15,14 @@ class ProductFormActivity : ComponentActivity() {
         setContent {
             AluveryTheme {
                 Surface {
-                    ProductFormScreen(onSaveClick = { product ->
-                        dao.save(product)
-                        finish()
-                    })
+                    ProductFormScreen(
+                        onSaveClick = { product ->
+                            dao.save(product)
+                            finish()
+                        }
+                    )
                 }
             }
-        }
-    }
-}
-
-@Composable
-fun ProductFormScreen(
-    onSaveClick: (Product) -> Unit = {}
-) {
-    Column(
-        Modifier
-            .fillMaxSize()
-            .padding(horizontal = 16.dp)
-            .verticalScroll(rememberScrollState()),
-        verticalArrangement = Arrangement.spacedBy(16.dp)
-    ) {
-        Spacer(modifier = Modifier)
-        Text(
-            text = stringResource(R.string.product_form_title),
-            Modifier.fillMaxWidth(),
-            fontSize = 28.sp
-        )
-
-        var url by remember {
-            mutableStateOf("")
-        }
-        if (url.isNotBlank()) {
-            AsyncImage(
-                model = url,
-                contentDescription = null,
-                Modifier
-                    .fillMaxWidth()
-                    .heightIn(200.dp),
-                contentScale = ContentScale.Crop,
-                placeholder = painterResource(id = R.drawable.placeholder),
-                error = painterResource(id = R.drawable.placeholder)
-            )
-        }
-        TextField(
-            value = url,
-            onValueChange = {
-                url = it
-            },
-            Modifier.fillMaxWidth(),
-            label = {
-                Text("Url of the image")
-            },
-            keyboardOptions = KeyboardOptions(
-                keyboardType = KeyboardType.Uri,
-                imeAction = ImeAction.Next
-            )
-        )
-
-        var name by remember {
-            mutableStateOf("")
-        }
-        TextField(
-            value = name,
-            onValueChange = {
-                name = it
-            },
-            Modifier.fillMaxWidth(),
-            label = {
-                Text("Name")
-            },
-            keyboardOptions = KeyboardOptions(
-                keyboardType = KeyboardType.Text,
-                imeAction = ImeAction.Next,
-                capitalization = KeyboardCapitalization.Words
-            )
-        )
-
-        var price by remember {
-            mutableStateOf("")
-        }
-        var isPriceError by remember {
-            mutableStateOf(false)
-        }
-
-        Column {
-            TextField(
-                value = price,
-                onValueChange = {
-                    isPriceError = try {
-                        BigDecimal(it)
-                        false
-                    } catch (e: IllegalArgumentException) {
-                        it.isNotEmpty()
-                    }
-                    price = it
-                },
-                Modifier.fillMaxWidth(),
-                isError = isPriceError,
-                label = {
-                    Text("Price")
-                },
-                keyboardOptions = KeyboardOptions(
-                    keyboardType = KeyboardType.Decimal,
-                    imeAction = ImeAction.Next
-                )
-            )
-            if (isPriceError) {
-                Text(
-                    text = "The price must be a decimal value",
-                    color = MaterialTheme.colors.error,
-                    style = MaterialTheme.typography.caption,
-                    modifier = Modifier.padding(start = 16.dp)
-                )
-            }
-        }
-        var description by remember {
-            mutableStateOf("")
-        }
-        TextField(
-            value = description,
-            onValueChange = {
-                description = it
-            },
-            Modifier.fillMaxWidth()
-                .heightIn(min = 100.dp),
-            label = {
-                Text("Description")
-            },
-            keyboardOptions = KeyboardOptions(
-                keyboardType = KeyboardType.Text,
-                capitalization = KeyboardCapitalization.Sentences
-            )
-        )
-
-        Button(onClick = {
-
-            val convertedPrice = try {
-                BigDecimal(price)
-            } catch (e: NumberFormatException) {
-                BigDecimal.ZERO
-            }
-            val product = Product(
-                name = name,
-                image = url,
-                price = convertedPrice,
-                description = description
-            )
-            onSaveClick(product)
-        }, Modifier.fillMaxWidth()) {
-            Text(text = "Save")
-        }
-        Spacer(modifier = Modifier)
-    }
-}
-
-@Composable
-@Preview
-fun ProductFormScreenPreview() {
-    AluveryTheme {
-        Surface {
-            ProductFormScreen()
         }
     }
 }
